apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.29.0
redis==5.0.8
python-dotenv==1.0.1
alembic==1.13.2
psycopg2-binary==2.9.9
